2.0.3: firmware-wide review - OOM aborts, lost writes, warning sweep
A full review pass over the firmware. Six parallel subsystem audits were
launched and all six died on an API session limit, so this is what I verified
myself: broad mechanical sweeps over the whole tree plus targeted reads of the
paths where a defect costs the user something.

Allocation failures that rebooted the device. Under -fno-exceptions a failing
bare new does not throw, it terminates the firmware, and the earlier nothrow
sweep missed four sites on paths a user reaches normally: the PNG scaling
accumulators (any cover that needs resizing), the file handles in the JPEG and
PNG decoders (any image inside a book), and the captive-portal DNS server
(allocated exactly when the heap is tightest). Each now degrades instead: no
cover, skipped image, portal without its redirect.

Atomic writes could lose a file outright. Both writers replace a file as
remove-then-rename, because SdFat will not rename over an existing name; a
power loss inside that window left the payload complete under "<path>.tmp" and
no canonical file, and nothing ever looked there again. Settings did not go one
save stale, they reset; a reading position was not slightly old, it was gone.
Reads now claim an orphaned .tmp on the first miss, and the config loaders -
which gate on exists() before reading - call the recovery explicitly first.

ctype calls on signed chars in the vendored PDF parser, three sites the
library's own "& 255" idiom had missed: a byte above 0x7F in a malformed hex
string indexes the ctype table out of range.

The compiler had never been asked. -Wall -Wextra are permanent build flags now,
and every warning they raised in our code is fixed: an unreachable strlen(NULL)
in the access-point path, two member-initializer orders that lied about
themselves, two precedence-ambiguous conditions (both already correct), four
format specifiers disagreeing with their arguments, a footer-counter buffer one
char short of its worst case, thirteen dead layout locals, and a memset over a
type with member initializers. A clean release build emits zero warnings from
our code.

Verified and deliberately unchanged: settings JSON clamps every value by type;
the OTA semver comparison; the reader's position restore clamping against a
re-paginated chapter; RenderLock ownership and the main-loop/render-task lock
order; the sleep-wallpaper ring buffer; keyboard entry; cppcheck at lowest
severity. 117/117 host tests, i18n validation, release at 89.8%.

Not verified on hardware: the device has been asleep since the 2.0.2 sleep test
and its USB port returns only after a physical power-button press.

+++ src/main.cpp
@@ -594,7 +594,7 @@ void loop() {
         RenderLock renderLock;
         const uint32_t bufferSize = display.getBufferSize();
         logSerial.setTxTimeoutMs(1000);
-        logSerial.printf("SCREENSHOT_START:%d\n", bufferSize);
+        logSerial.printf("SCREENSHOT_START:%u\n", (unsigned)bufferSize);
         uint8_t* buf = display.getFrameBuffer();
         uint32_t bytesSent = 0;
         while (bytesSent < bufferSize) {
